+++ Anarchy/ChangeLog.md
@@ -1,4 +1,5 @@
 ﻿## Patch V1.4.3
 * Fixed tooltip related issues.
 * Updated localization. Added I18n Everywhere dependency.
+* Fixed Flaming chirper with disabled radio.
 *